Creates endzone test scene static endzone
The static endzone object in the endzone test scene will only print a debug message if the endzone is triggered by a collision object with the "Player" GameObject tag; this trigger assumes that the player object itself has a collider component, not one of its child objects. A tag has been added to the player objects in the EndZoneTest scene. The endzone still needs adjustable width and height.