Game state manager and game manager
First iteration of game manager and game state manager.

The game manager will be the first thing that the game loads, it is a singleton and it will be in charge of handling the menus and level transitions.

In this game manager we will instantiate and setup all the other independent systems (Inventory, Timer, Level Generator, Game State Manager, Player Controllers).

TODO: Right now it starts the game right away, once we have the menus then we need to connect the ui buttons via events to either exit the game (Close everything), show credits (Show a little lovely message from the dev team), start game (Send a OnGameStart event).

Mocks were added to avoid compilation errors and to give a sense of the wiring we should do.

GameStateManager will be the game object in charge of managing the game loop inside a single level, it's only job is to react to events from other systems (i.e. timer, inventory) and update the game state machine.

It sends event when the level objectives are completed, when the player fails and when the player succeeds