Adds light to end zone
A light 2D component is added to end zone. Its radius is adjusted to match the max dimension of the end zone. In addition, the player tag has been applied to the "visual" child objects in the players, as these contain the colliders for triggering the end zone. The end zone has also been turned into a prefab, although a temp sprite is still being used. The light 2D also requires a global light to be in the screen, and this has been added to the EndZoneTest scene.