Flags players in end zone
The GameStateManager has a new bool and int value to keep track of when players are in an end zone. The bool _exitCondition is set to true when a player triggers an end zone, and is set to false when the last player in an end zone exits said end zone. If _exitCondition is true when the players have obtained all mandatory items, the EndGameSuccessState() is immediately called if at least 1 player is also on an end zone. The end zone light cue does not follow this logic yet since the inventory still needs to be implemented to this branch.