Moves logic from game state manager to end zone
Instead of keeping track of the amount of players that remains in the end zone in the game state manager class, we now do that in the end zone and just notify whenever a player _enters_ or when no more players are inside an end zone.

Colliders for the end zone are now deactivated on start and will only be activated once the main objective has been completed, thus avoiding sending extra events that would basically do nothing on the game state manager as the win condition will always make sure that the level can't end in that case.

Whenever the end zone is empty then we deactivate the exit condition