Utilizes OverlapCollider to more accurately calculate players in the end zone
When a player exits the end zone, a check is done in FixedUpdate to see if any colliders are still overlapping with the end zone trigger. If not, the callback function in OnEndZoneEmpty is invoked. This might be altered if items will be throwable/droppable, requiring the colliders in the zoneColliders array to check if there are any more zoneColliders with the GameObject "Player" tag. The size of the results array would also have to be adjusted, or the contactFilter would need to filter for only player objects (I'm not familiar with ContactFilters, though, so clarification on that would be appreciated).