First level and some wiring
So here's a non working first iteration of a level, I went for tile map with both the floors and the walls, it's not quite right but it dramatically increases the development of a level at the cost of a poopy level.

Having issues with the sorting layers on the rendering, I'm not quite sure right now on how to make it so that the lower half of the body hides behind walls. If someone has some more experience with that, I would greatly appreciate your help with it.

Tomorrow, if we're tight, let's focus on a single level and make the shop circle back to the same level I guess haha.

 The game manager is now the one in charge of transitioning from level to level and show/hide UIs

Pending: Wiring state, UI w/timer, end zones, inventory and objectives, connect the audio.