Completes UI functionality for shop
Uses MockInventory as a inventory for the coins and buff/debuff objects. When a level is completed, the ShopUI is activated, and the player can purchase items they missed in the level (identified in the MockInventory script in the Mock GameObject). The purchased items still need to be returned to the MockInventory after continuing from the shop.